Adds sepcial case to my custom cut and copy functions.
If the region is inactive, operate on the current line.
When copying, also advance to the next line, and set this-command to
kill-region so that a repeat of the same command will append the kill ring
entry.
Repeated cuts (kills) of the current line append the kill ring without that
extra work.

Clunky: The extra copy I make to a register does not share the behavior of
repeated commands.